🛠 Transferred get_session to dependencies (#36)
